pyproject.toml
[tool.black]
exclude = '''/(\.eggs|\.git|\.hg\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist|migrations)/'''
line-length = 88

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.poetry]
name = "unisport"
version = "0.1.0"
description = "Unisport product service"
authors = ["simon_pears"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0"
djangorestframework = "^3.13.1"
djangorestframework-jsonapi = "^5.0.0"
flake8 = "^5.0.4"
django-debug-toolbar = "^3.5.0"
drf-spectacular = "^0.23.1"
httpx = "^0.23.0"
pydantic = "^1.9.1"
django-structlog = "^3.0.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
black = "^22.6.0"
isort = "^5.10.1"
PyHamcrest = "^2.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
